remove da bad tings
