+++ .vscode/settings.json
@@ -6,5 +6,6 @@
   "eslint.validate": ["typescript"],
   "[typescript]": {
     "editor.defaultFormatter": "dbaeumer.vscode-eslint"
-  }
+  },
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
