+++ scripts/build-mcpb-schema.js
@@ -1,9 +1,7 @@
 import { McpbManifestSchema as McpbManifestSchema_v0_1 } from "../dist/schemas/0.1.js";
 import { McpbManifestSchema as McpbManifestSchema_v0_2 } from "../dist/schemas/0.2.js";
-import {
-  McpbManifestSchema as McpbManifestSchema_v0_3,
-  McpbSignatureInfoSchema,
-} from "../dist/schemas/0.3.js";
+import { McpbManifestSchema as McpbManifestSchema_v0_3 } from "../dist/schemas/0.3.js";
+import { McpbSignatureInfoSchema } from "../dist/shared/common.js";
 import { zodToJsonSchema } from "zod-to-json-schema";
 import fs from "node:fs/promises";
 import path from "node:path";
