+++ src/browser.ts
@@ -2,4 +2,3 @@
 export * from "./schemas/index.js";
 export * from "./shared/config.js";
 export * from "./shared/constants.js";
-export * from "./types.js";
