+++ src/cli.ts
@@ -6,7 +6,6 @@ export * from "./cli/pack.js";
 export * from "./schemas/index.js";
 export * from "./shared/config.js";
 export * from "./shared/constants.js";
-export * from "./types.js";
 
 // Include node exports since CLI needs them
 export * from "./node/files.js";
