+++ src/index.ts
@@ -8,4 +8,3 @@ export * from "./node/validate.js";
 export * from "./schemas/index.js";
 export * from "./shared/config.js";
 export * from "./shared/constants.js";
-export * from "./types.js";
