+++ src/node.ts
@@ -7,4 +7,3 @@ export * from "./node/validate.js";
 export * from "./schemas/index.js";
 export * from "./shared/config.js";
 export * from "./shared/constants.js";
-export * from "./types.js";
