+++ src/node/sign.ts
@@ -5,8 +5,9 @@ import forge from "node-forge";
 import { tmpdir } from "os";
 import { join } from "path";
 import { promisify } from "util";
+import type { z } from "zod";
 
-import type { McpbSignatureInfo } from "../types.js";
+import type { McpbSignatureInfoSchema } from "../shared/common.js";
 
 // Signature block markers
 const SIGNATURE_HEADER = "MCPB_SIG_V1";
@@ -101,7 +102,7 @@ export function signMcpbFile(
  */
 export async function verifyMcpbFile(
   mcpbPath: string,
-): Promise<McpbSignatureInfo> {
+): Promise<z.infer<typeof McpbSignatureInfoSchema>> {
   try {
     const fileContent = readFileSync(mcpbPath);
