+++ src/schemas/0.1.ts
@@ -71,11 +71,6 @@ export const McpbUserConfigurationOptionSchema = z.strictObject({
   max: z.number().optional(),
 });
 
-export const McpbUserConfigValuesSchema = z.record(
-  z.string(),
-  z.union([z.string(), z.number(), z.boolean(), z.array(z.string())]),
-);
-
 export const McpbManifestSchema = z
   .strictObject({
     $schema: z.string().optional(),
@@ -113,12 +108,3 @@ export const McpbManifestSchema = z
     message:
       "Either 'dxt_version' (deprecated) or 'manifest_version' must be provided",
   });
-
-export const McpbSignatureInfoSchema = z.strictObject({
-  status: z.enum(["signed", "unsigned", "self-signed"]),
-  publisher: z.string().optional(),
-  issuer: z.string().optional(),
-  valid_from: z.string().optional(),
-  valid_to: z.string().optional(),
-  fingerprint: z.string().optional(),
-});
