+++ src/schemas_loose/0.3.ts
@@ -77,11 +77,6 @@ export const McpbUserConfigurationOptionSchema = z.object({
   max: z.number().optional(),
 });
 
-export const McpbUserConfigValuesSchema = z.record(
-  z.string(),
-  z.union([z.string(), z.number(), z.boolean(), z.array(z.string())]),
-);
-
 export const McpbManifestLocalizationSchema = z
   .object({
     resources: z
@@ -153,12 +148,3 @@ export const McpbManifestSchema = z
     message:
       "Either 'dxt_version' (deprecated) or 'manifest_version' must be provided",
   });
-
-export const McpbSignatureInfoSchema = z.object({
-  status: z.enum(["signed", "unsigned", "self-signed"]),
-  publisher: z.string().optional(),
-  issuer: z.string().optional(),
-  valid_from: z.string().optional(),
-  valid_to: z.string().optional(),
-  fingerprint: z.string().optional(),
-});
