+++ src/shared/common.ts
@@ -1,0 +1,15 @@
+import { z } from "zod";
+
+export const McpbUserConfigValuesSchema = z.record(
+  z.string(),
+  z.union([z.string(), z.number(), z.boolean(), z.array(z.string())]),
+);
+
+export const McpbSignatureInfoSchema = z.strictObject({
+  status: z.enum(["signed", "unsigned", "self-signed"]),
+  publisher: z.string().optional(),
+  issuer: z.string().optional(),
+  valid_from: z.string().optional(),
+  valid_to: z.string().optional(),
+  fingerprint: z.string().optional(),
+});
